feat: rust ceramic tests (#2965)
* feat: rust ceramic tests

* fix: update test framework to work with Rust IPFS

This change makes it possible to call createIPFS to get a Rust IPFS
implementation on a random port and that can be stopped. This allows
most tests to pass.

Additionally a few tests needed to be updated with their delay to handle
subtle differences in how Go vs Rust do async work. Specifically Go
seems to be faster a learning that newly connected peers are also
interested in the same gossipsub topics. Rust needs a small delay after
telling peers to subscribe before publishing otherwise publish will fail
with an error about the peer not having sufficient peers.

* fix: run prettier

Also adds format command to package.json so its easy to run

* fix: add disposable doc comment

---------

Co-authored-by: Nathaniel Cook <nvcook42@gmail.com>