fix: adjust capability tests (correctly this time)
use lowercase DID address in tests. address is mixed case on wallet with eip55 but dids@5.0.2 returns lowercase as we previously relied on that in ceramic.

our tests assigned our wallet value directly in some cases, skipping the value from the verified DID.
`ModelInstanceDocument._makeRawGenesis` accepts the `metadata.controller` without changing it, so
some tests returned uppercase, and others lowercase from (signer.did).
we could consider validating the format of the passed in controller when creating the genesis commit.