chore: rely on more ceramic one defaults and use stdout/err from parent
We bind all IPs on the host so it works in docker and generally as expected.
This feels a bit janky and not in one place still. We really don't pipe down rust options, so it mostly relies on env vars and tries to keep the behavior the same for tests with temp dirs and inmemory, running from the cli, or running in docker