chore: (breaking changes) use rust "ipfs" by default (#3250)
* chore: change default rust "ipfs" rpc port to 5101

* chore: replace CERAMIC_RECON_MODE with IPFS_FLAVOR and access from EnvironmentUtils

we no longer have rust ceramic without recon mode, so we don't need both

* chore: update tests to use common recon/v4 setting

* chore: prettier

* chore: rely on more ceramic one defaults and use stdout/err from parent

We bind all IPs on the host so it works in docker and generally as expected.
This feels a bit janky and not in one place still. We really don't pipe down rust options, so it mostly relies on env vars and tries to keep the behavior the same for tests with temp dirs and inmemory, running from the cli, or running in docker

* chore: set path to rust-ceramic vars in dockerfile

can be overriden via `docker run -e CERAMIC_ONE_PATH=... -e CERAMIC_ONE_STORE_DIR=...`

* chore: remove CERAMIC_RECON_MODE from docs/test set up

* chore: fix lint error and unnecessary switch branch

* chore: use default IPFS config in tests

we only override a portion which was invalid, now we just use env vars/default rust

* chore: override metrics and swarm addresses for tests so they pass

this has the unfortunate side effect of using a temp directory if network or store env vars aren't set that gets cleaned up, but fixing that will require making the tests explicit about what they want. should happen next though.

* chore: specify CERAMIC_ONE_* variables in both layers of dockerfile

we sometimes run the js-ceramic layer directly and this makes it start up without external config